// src/renderer/tools/animation-editor/tools/types.ts
export enum SyncType {
    /**
     * Defines the animation to be played alone.
     */
    Animation = "Animation",
    /**
     * Defines the animation to be played with the other ones of the object.
     */
    Object = "Object",
    /**
     * Defines the animation to be played with all other animations in the scene.
     */
    Scene = "Scene",
}

export enum SelectedTab {
    /**
     * Defines the Id of the timelien tab.
     */
    Timeline = "Timeline",
    /**
     * Defines the Id of the graph tab.
     */
    Graph = "Graph",
}